+++ src/leaderboard/components/LeaderPage/style.less
@@ -1,30 +1,87 @@
 @import '../../../styles/vars.less';
 
 .leader-page {
-  .title {
+  &-top {
+    align-items: center;
+    display: flex;
+    justify-content: space-between;
+    width: 90%;
+  }
+
+  &-title {
     display: inline-block;
     font-size: @general-heading-1;
     font-weight: 700;
   }
 
-  .ant-dropdown-menu {
-    width: fit-content;
-  }
-
-  .ant-dropdown-link {
+  &-dropdown-link {
     color: black;
     float: right;
     font-size: 1.5rem;
     font-weight: 700;
-    padding: 1rem 0;
+    margin: 1rem 0;
 
     svg {
       margin-left: 5px;
     }
   }
+
+  &-timeframe {
+    color: black;
+    font-size: 1rem;
+  }
+
+  &-leaderboard {
+    > * {
+      max-width: 1700px;
+      width: 90%;
+    }
+
+    .top-three {
+      align-items: center;
+      display: flex;
+      flex-wrap: wrap;
+      justify-content: space-between;
+      margin: 20px 0;
+
+      > * {
+        margin: 10px 0;
+      }
+    }
+
+    .four-and-more {
+      border-radius: 10px;
+      overflow: hidden;
+    }
+  }
 }
 
-.leader-timeframe {
-  color: black;
-  font-size: 1rem;
+@media screen and (max-width: 1305px) {
+  .leader-page {
+    &-leaderboard {
+      .top-three {
+        justify-content: space-evenly;
+      }
+    }
+  }
+}
+
+@media screen and (max-width: 958px) {
+  .leader-page {
+    &-top {
+      align-items: flex-start;
+      flex-direction: column;
+    }
+
+    &-leaderboard {
+      > * {
+        width: 100%;
+      }
+
+      .top-three {
+        justify-content: center;
+        margin: 0 0 20px 0;
+      }
+    }
+  }
 }
