--- src/leaderboard/components/Leaderboard/index.tsx
@@ -1,21 +1,0 @@
-import React from 'react';
-
-import TopThree from '../../containers/TopThree';
-import FourAndMore from '../../containers/FourAndMore';
-
-import './style.less';
-
-const Leaderboard: React.FC = () => {
-  return (
-    <div className="leaderboard">
-      <div className="top-three">
-        <TopThree />
-      </div>
-      <div className="four-and-more">
-        <FourAndMore />
-      </div>
-    </div>
-  );
-};
-
-export default Leaderboard;
