--- src/leaderboard/components/Leaderboard/style.less
@@ -1,43 +1,0 @@
-.leaderboard {
-  > * {
-    max-width: 1700px;
-    width: 90%;
-  }
-
-  .top-three {
-    align-items: center;
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: space-between;
-    margin: 20px 0;
-
-    > * {
-      margin: 10px 0;
-    }
-  }
-
-  .four-and-more {
-    border-radius: 10px;
-    overflow: hidden;
-  }
-}
-
-@media screen and (max-width: 1305px) {
-  .leaderboard {
-    .top-three {
-      justify-content: space-evenly;
-    }
-  }
-}
-
-@media screen and (max-width: 958px) {
-  .leaderboard {
-    > * {
-      width: 100%;
-    }
-
-    .top-three {
-      justify-content: center;
-    }
-  }
-}
