+++ src/leaderboard/leaderboardActions.ts
@@ -1,18 +1,12 @@
-import { CLEAR_LEADERBOARD, FETCH_LEADERBOARD, LEADERBOARD_ERROR, UPDATE_TIMEFRAME, ThunkActionCreator } from './leaderboardTypes';
+/* eslint-disable import/prefer-default-export */
+import { FETCH_LEADERBOARD, LEADERBOARD_ERROR, ThunkActionCreator } from './leaderboardTypes';
 import { logoutUser } from '../auth/authActions';
 
 import Config from '../config';
 import { fetchService } from '../utils';
 
-export const fetchLeaderboard: ThunkActionCreator = (offset: number = 0, limit: number, from?: number, to?: number, resetUsers?: boolean) => async (
-  dispatch,
-) => {
+export const fetchLeaderboard: ThunkActionCreator = (offset: number = 0, limit: number, from?: number, to?: number) => async (dispatch) => {
   try {
-    if (resetUsers) {
-      dispatch({
-        type: CLEAR_LEADERBOARD,
-      });
-    }
     const url = `${Config.API_URL}${Config.routes.leaderboard}?offset=${offset}&limit=${limit}${from ? `&from=${from}` : ''}${to ? `&to=${to}` : ''}`;
     const data = await fetchService(url, 'GET', 'json', {
       requiresAuthorization: true,
@@ -33,11 +27,4 @@ export const fetchLeaderboard: ThunkActionCreator = (offset: number = 0, limit: 
       payload: error,
     });
   }
-};
-
-export const updateTimeframe: ThunkActionCreator = (timeframe) => (dispatch) => {
-  dispatch({
-    type: UPDATE_TIMEFRAME,
-    payload: timeframe,
-  });
 };
